Learning to match in public policy applications
Learning to match: methods and challenges in the context of public policy applications. 

Well known applications of machine learning, such as document retrieval, data integration or recommender systems can be categorized as “learning to match” models. 
These matching algorithms are increasingly being applied 
outside of the Computer Science and Engineering realm to different policy problems such as; hospital-bed assignments, organ transplants, refugee allocation to different locations, active unemployment policies, among others.

In this project we use a data base on characteristics of European soccer teams and players to analyze the methodological challenges of learning to match models when applied to policy problems.